# api.yaml
openapi: "3.0.0"
info:
  title: Otto user service
  version: 1.0
paths:
...
  /reshare:
    post:
      operationId: createAReshare
      summary: create a reshare
      requestBody:
        description: reshare to create
        required: true
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/NewReshare"
      responses:
        '201':
          description: |-
            201 response
  /rehare/{uuid}:
    get:
      operationId: getReshare
      summary: get a reshare
      parameters:
        - in: path
          name: uuid
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: get a reshare
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Reshare'
        '404':
          description: |-
            404 response
...
components:
  schemas:
    User:
      type: object
      required:
        - uuid
        - name
        - created
      properties:
        uuid:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        follows:
          type: array
          items:
            $ref: "#/components/schemas/Follow"
    Follow:
      type: object
      required:
        - uuid
        - user
        - following
      properties:
        uuid:
          type: string
          format: uuid
        user:
          type: object
          $ref: "#/components/schemas/User"
        following:
          type: object
          $ref: "#/components/schemas/User"
...
    Reshare:
      type: object
      required:
        - uuid
        - message
        - reshare
      properties:
        uuid:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        message:
          type: object
          $ref: "#/components/schemas/Message"
        reshare:
          type: object
          $ref: "#/components/schemas/Message"
    NewReshare:
      type: object
      properties:
        message:
          type: object
          $ref: "#/components/schemas/NewMessage"
        reshare:
          type: object
          $ref: "#/components/schemas/Message"
    Message:
      type: object
      required:
        - uuid
        - created
        - text
        - user
        - visibility
      properties:
        uuid:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        text:
          type: string
        user:
          type: object
          $ref: "#/components/schemas/User"
        visibility:
          type: string
          $ref: "#/components/schemas/Visibility"
    NewMessage:
      type: object
      required:
        - message
        - user
      properties:
        message:
          type: string
        user:
          type: object
          $ref: "#/components/schemas/User"
...
    Visibility:
      type: string
      enum:
        - public
        - limited
        - private