# internal/model/message.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"github.com/danielmunro/otto-community-service/internal/enum"
	"time"
)

type Message struct {
	Uuid string `json:"uuid"`

	CreatedAt time.Time `json:"created_at"`

	Text string `json:"text"`

	User User `json:"user"`

	Visibility enum.Visibility `json:"visibility"`
}