# internal/model/new_message.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

type NewMessage struct {
	Text string `json:"text"`

	User User `json:"user"`
}