# internal/model/reshare.go
/*
 * Otto user service
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package model

import (
	"time"
)

type Reshare struct {
	Uuid string `json:"uuid"`

	CreatedAt time.Time `json:"created_at"`

	Message Message `json:"message"`

	Reshare Message `json:"reshare"`
}